backend/pyproject.toml
[project]
name = "nuxt-notes-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "websockets>=13.0",
    "claude-agent-sdk>=0.0.14",
    "python-dotenv>=1.0.0",
    "supabase>=2.28.3",
]
